Clean up build and test output for Gentoo.

+++ test/example/cfg.c
@@ -4,7 +4,7 @@ int main(int argc, char **argv)
 {
     if(argc & 1) {
         for(int i = 0; i < argc; i++) {
-            printf("%d\n", argv[i]);
+            printf("%ld\n", (long)argv[i]);
         }
     }
