+++ test/example/detectnullfail.c
@@ -3,6 +3,6 @@
 
 void egalito_null_ptr_check_fail(void) {
     printf("egalito null ptr check failed at 0x%lx\n",
-        __builtin_return_address(0));
+        (unsigned long)__builtin_return_address(0));
     exit(1);
 }
