+++ test/example/getenv.c
@@ -2,7 +2,7 @@
 #include <stdlib.h>
 
 int main(int argc, char **argv, char **envp) {
-    printf("note: envp is %lx\n", envp);
+    printf("note: envp is %lx\n", (unsigned long)envp);
     printf("note: *envp is %s\n", *envp);
     printf("getenv: PATH=%s\n", getenv("PATH"));
     puts("iterating through envp manually:");
